backend/requirements.txt
# HiRA Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LLM Providers
anthropic==0.18.1
openai==1.10.0

# Vector Database
chromadb==0.4.22
numpy<2.0  # ChromaDB 0.4.22 requires NumPy 1.x

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Utilities
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Zoom Bot Integration
deepgram-sdk==3.0.0
websockets==12.0
requests==2.31.0
elevenlabs==0.2.27
pydub==0.25.1
